+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v4.62.0 Features
 
+- `[Datepicker]` Added legend load for datepicker. ([NG#1261](https://github.com/infor-design/enterprise-ng/issues/1261))
 - `[File Upload]` Added setFailed status ([#5671](https://github.com/infor-design/enterprise/issues/5671))
 - `[Icon]` Created a puppeteer script for the new launch icon. ([#5854](https://github.com/infor-design/enterprise/issues/5854))
 - `[Icon]` Created a puppeteer script for the new mobile icon. ([#6199](https://github.com/infor-design/enterprise/issues/6199))